bugfixes
- Fixed a massive rotation bug due to some network optimizations done previously.
- Fixed a potential null pointer crash

+++ src/main/java/train/common/core/network/PacketRollingStockRotation.java
@@ -38,10 +38,10 @@ public class PacketRollingStockRotation implements IMessage {
 	public void fromBytes(ByteBuf bbuf) {
 
 		this.entityID = bbuf.readInt();
-		this.rotationYawServer = bbuf.readInt();
-		this.realRotation = bbuf.readInt();
+		this.rotationYawServer = bbuf.readFloat();
+		this.realRotation = bbuf.readFloat();
 		this.anglePitch = bbuf.readInt();
-		this.posY = bbuf.readInt();
+		this.posY = bbuf.readFloat();
 		this.isInReverse = bbuf.readBoolean();
 	}
